Rename core to gateway
