--- core/src/main/java/dev/gihwan/tollgate/core/TollgateBuilder.java
@@ -1,68 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 - 2021 Gihwan Kim
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package dev.gihwan.tollgate.core;
-
-import com.linecorp.armeria.server.Server;
-import com.linecorp.armeria.server.ServerBuilder;
-import com.linecorp.armeria.server.ServiceBindingBuilder;
-import com.linecorp.armeria.server.healthcheck.HealthCheckService;
-
-public abstract class TollgateBuilder {
-
-    private final ServerBuilder serverBuilder = Server.builder();
-
-    protected TollgateBuilder() {}
-
-    public abstract UpstreamBindingBuilder<? extends TollgateBuilder> route();
-
-    /**
-     * @see ServerBuilder#http(int)
-     */
-    public TollgateBuilder http(int port) {
-        serverBuilder.http(port);
-        return this;
-    }
-
-    public TollgateBuilder healthCheck(String healthCheckPath) {
-        return healthCheck(healthCheckPath, HealthCheckService.of());
-    }
-
-    public TollgateBuilder healthCheck(String healthCheckPath, HealthCheckService healthCheckService) {
-        serverBuilder.service(healthCheckPath, healthCheckService);
-        return this;
-    }
-
-    public TollgateBuilder upstream(String pathPattern, Upstream upstream) {
-        return route().path(pathPattern).build(upstream);
-    }
-
-    public final Tollgate build() {
-        return new Tollgate(serverBuilder.build());
-    }
-
-    protected final ServiceBindingBuilder serverRoute() {
-        return serverBuilder.route();
-    }
-}
